+++ spp_cr_type_assign_program/tests/__init__.py
@@ -1,1 +1,2 @@
 from . import test_assign_program
+from . import test_program_access